Message log fills/scrolls bottom up, message sending
Tweaked the chat log so that messages appear at the
end of the chat log. Needed to use a dirty hack where
I make a copy to reverse the list of messages, as
`justify-end` decided it wanted to break the scrolling.

Added text input for user texting. Pressing enter
"sends" the message by adding it to the end of
the message log. Svelte's reactivity takes care
of re-grouping chat messages and everything, because...

SVELTE IS FREAKING AWESOME!!!